pyproject.toml
[project]
name = "project-run"
version = "0.1.0"
description = "Run tracker backend"
authors = [
    {name = "Denis"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "haversine (>=2.9.0,<3.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "black (>=25.12.0,<26.0.0)"
]

[tool.poetry]
name = "project-run"
version = "0.1.0"
description = ""
authors = ["denchickuss@yandex.ru"]

package-mode = false

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    "migrations",
    "__pycache__",
    "config/settings",
]


[tool.black]
line-length = 88
target-version = ["py312"]
exclude = "migrations"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]



